A36224-000 Version 0.0.5
Started to Test ADCs. Wasn't working at first, because when defining the
skips, we need to skip all of the unused inputs, not just the ones that
are enabled and not used. So far, it seems like I'm getting something on
the input, but it seems a bit low. For an input of 2.04 volts, I'm
getting the 12 bit value 0x0614. When referenced to 5V, this is more
like 1.899V. This will have to be fixed.