Version 0.0.6
Debugged cabinet and coolant temperature calibrations. Also did
troubleshooting for ADC problem, where voltage measured was low. It
turned out the problem was on the board- there was a voltage coming into
the ADC that was outside of its range. The differential amplifiers for
Analog inputs 2 and 3 were floating, and sending a slightly negative
voltage.