A36224-000 Version 0.0.7
I didn't actually change anything in the code. Currently working on
testing the SF6 pressure input. For some reason, the scaling on the
unused external scale is causing an integer overflow. The math makes
sense- it should cause an overflow, but I'm not sure what the analog
module is trying to do. Next step is to update the analog module. Next
commit should include changes to the project to include github-synced
versions of the ETM modules.