A36224-000 Version 0.0.8
Configured the project to run on ETM Modules in another folder. This
folder is linked to github for updates. I checked the SF6 pressure
scaling. It works, but I will have to figure out how to deal with a
problem if the input is less than the offset (this would cause a
negative number stored as an unsigned int).