A36224-000 Version 0.0.10.
Set up the DoA36224 to work on the 10ms timer. Started working on the
SF6 pressure control. It is not working yet. I am unable to set the
digital bits controlling the solenoid relay from that function. I will
fix that on Monday. Also, the coolant temperature sensor seems to be
behaving strangely, but I suspect that there is a resistor on the board
that should not be there. (probably have to remove R67. I thought I did
this, but I'll check later).

This code does compile, and should work fine. The sf6 controls just dont
do anything just yet.