A36224-000 Version 0.0.11
Set up SF6 pump control. So far, it is tested and working well. Next
steps are to figure out how to store the bottle status, and implement
override and reset commands on the ECB and gui. A strange thing is
happening with Analog 0. the input is tied to ground, but it is seeing
1.33V. Probably on the board, but I'm not sure how or why yet.